fix: remove useless folder
