+++ src/strategy/strategy/coach/penalty.py
@@ -1,7 +1,7 @@
 
 from strategy.behaviour import LeafNode, Selector, Sequence, TaskStatus
 from strategy.blackboard import Blackboard
-from strategy.robots.kickoff.our_kick_off.goalkeeper import TheirActionGoalKeeper
+from strategy.robots.kickoff.goalkeeper import TheirActionGoalKeeper
 from strategy.robots.penalty.our_penalty.attacker import OurActionAttacker, TheirActionAttacker
