+++ src/strategy/strategy/robots/kickoff/defender.py
