+++ src/strategy/strategy/robots/kickoff/goalkeeper.py
